+++ packages/components-react/src/examples/CanvasExample.tsx
@@ -1,11 +1,30 @@
-import { PCanvas, PText } from '@porsche-design-system/components-react';
+import {
+  PCanvas,
+  PText,
+  PButton,
+  type CanvasSidebarStartUpdateEventDetail,
+} from '@porsche-design-system/components-react';
+import { breakpointS } from '@porsche-design-system/components-react/styles';
+import { useCallback, useState } from 'react';
 
 export const CanvasExamplePage = (): JSX.Element => {
-  const style = `
-      body {
-        overflow-x: hidden;
-      }
+  const [isSidebarStartOpen, setIsSidebarStartOpen] = useState<boolean>(
+    // initially, sidebar should be closed on mobile and opened on desktop
+    window.matchMedia(`(min-width: ${breakpointS}px)`).matches
+  );
+  const [isSidebarEndOpen, setIsSidebarEndOpen] = useState<boolean>(false);
 
+  const onSidebarStartUpdate = useCallback((e: CustomEvent<CanvasSidebarStartUpdateEventDetail>) => {
+    setIsSidebarStartOpen(e.detail.open);
+  }, []);
+  const onSidebarEndOpen = useCallback(() => {
+    setIsSidebarEndOpen(true);
+  }, []);
+  const onSidebarEndDismiss = useCallback(() => {
+    setIsSidebarEndOpen(false);
+  }, []);
+
+  const style = `
       .-col-span-full-1 {
         grid-column: 1 / -1;
       }
@@ -35,26 +54,42 @@ export const CanvasExamplePage = (): JSX.Element => {
   return (
     <>
       <style dangerouslySetInnerHTML={{ __html: style }} />
-      <PCanvas>
+      <PCanvas
+        sidebarStartOpen={isSidebarStartOpen}
+        sidebarEndOpen={isSidebarEndOpen}
+        onSidebarStartUpdate={onSidebarStartUpdate}
+        onSidebarEndDismiss={onSidebarEndDismiss}
+      >
         <a slot="title" href="#">
           App Name
         </a>
 
-        <PText className="-col-span-full-1">Content</PText>
+        <PButton
+          slot="header-end"
+          icon="configurate"
+          variant="ghost"
+          compact={true}
+          hideLabel={true}
+          onClick={onSidebarEndOpen}
+        >
+          Open sidebar
+        </PButton>
 
-        <div className="tile -col-span-4">Grid span 4x</div>
-        <div className="tile -col-span-4">Grid span 4x</div>
-        <div className="tile -col-span-4">Grid span 4x</div>
+        <div className="-p-canvas-grid">
+          <PText className="-col-span-full-1">Content</PText>
 
-        <div className="tile -col-span-full-1">12 Grid columns</div>
-        <div className="tile -col-span-full-2">10 Grid columns</div>
-        <div className="tile -col-span-full-3">8 Grid columns</div>
+          <div className="tile -col-span-4">Grid span 4x</div>
+          <div className="tile -col-span-4">Grid span 4x</div>
+          <div className="tile -col-span-4">Grid span 4x</div>
 
-        <PText slot="footer" className="-col-span-full-1">
-          Footer
-        </PText>
-        <div slot="footer" className="tile -col-span-full-1">
-          12 Grid columns
+          <div className="tile -col-span-full-1">12 Grid columns</div>
+          <div className="tile -col-span-full-2">10 Grid columns</div>
+          <div className="tile -col-span-full-3">8 Grid columns</div>
+        </div>
+
+        <div slot="footer" className="-p-canvas-grid">
+          <PText className="-col-span-full-1">Footer</PText>
+          <div className="tile -col-span-full-1">12 Grid columns</div>
         </div>
 
         <div slot="sidebar-start">