+++ packages/components-vue/src/examples/CanvasExample.vue
@@ -1,23 +1,66 @@
 <script setup lang="ts">
-  import { PCanvas, PText } from '@porsche-design-system/components-vue';
+  import {
+    PCanvas,
+    PText,
+    PButton,
+    type CanvasSidebarStartUpdateEventDetail,
+  } from '@porsche-design-system/components-vue';
+  import { breakpointS } from '@porsche-design-system/components-vue/styles';
+  import { ref } from 'vue';
+
+  // initially, sidebar should be closed on mobile and opened on desktop
+  const isSidebarStartOpen = ref(window.matchMedia(`(min-width: ${breakpointS}px)`).matches);
+  const isSidebarEndOpen = ref(false);
+
+  const onSidebarStartUpdate = (e: CanvasSidebarStartUpdateEventDetail): void => {
+    isSidebarStartOpen.value = e.open;
+  };
+
+  const onSidebarEndOpen = (): void => {
+    isSidebarEndOpen.value = true;
+  };
+
+  const onSidebarEndDismiss = (): void => {
+    isSidebarEndOpen.value = false;
+  };
 </script>
 
 <template>
-  <PCanvas>
+  <PCanvas
+    :sidebarStartOpen="isSidebarStartOpen"
+    :sidebarEndOpen="isSidebarEndOpen"
+    @sidebarStartUpdate="onSidebarStartUpdate"
+    @sidebarEndDismiss="onSidebarEndDismiss"
+  >
     <a slot="title" href="#">App Name</a>
 
-    <PText class="-col-span-full-1">Content</PText>
+    <PButton
+      slot="header-end"
+      icon="configurate"
+      variant="ghost"
+      :compact="true"
+      :hideLabel="true"
+      @click="onSidebarEndOpen"
+    >
+      Open sidebar
+    </PButton>
 
-    <div class="tile -col-span-4">Grid span 4x</div>
-    <div class="tile -col-span-4">Grid span 4x</div>
-    <div class="tile -col-span-4">Grid span 4x</div>
+    <div class="-p-canvas-grid">
+      <PText class="-col-span-full-1">Content</PText>
 
-    <div class="tile -col-span-full-1">12 Grid columns</div>
-    <div class="tile -col-span-full-2">10 Grid columns</div>
-    <div class="tile -col-span-full-3">8 Grid columns</div>
+      <div class="tile -col-span-4">Grid span 4x</div>
+      <div class="tile -col-span-4">Grid span 4x</div>
+      <div class="tile -col-span-4">Grid span 4x</div>
 
-    <PText slot="footer" class="-col-span-full-1">Footer</PText>
-    <div slot="footer" class="tile -col-span-full-1">12 Grid columns</div>
+      <div class="tile -col-span-full-1">12 Grid columns</div>
+      <div class="tile -col-span-full-2">10 Grid columns</div>
+      <div class="tile -col-span-full-3">8 Grid columns</div>
+    </div>
+
+    <div slot="footer" class="-p-canvas-grid">
+      <PText class="-col-span-full-1">Footer</PText>
+      <div class="tile -col-span-full-1">12 Grid columns</div>
+    </div>
 
     <div slot="sidebar-start">
       <PText>Sidebar Start</PText>
@@ -27,12 +70,6 @@
     </div>
   </PCanvas>
 </template>
-
-<style>
-  body {
-    overflow-x: hidden;
-  }
-</style>
 
 <style scoped>
   .-col-span-full-1 {