+++ packages/components/scripts/wrapper-generator/UXPinReactWrapperGenerator.ts
@@ -14,8 +14,6 @@ const addNestedIndentation = (x: string): string => `  ${x}`;
 export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
   protected projectDir = 'uxpin-wrapper';
 
-  private spacingProps: string[] = ['top', 'left', 'right', 'bottom'].map((x) => `spacing${pascalCase(x)}`);
-
   constructor() {
     super();
     this.ignoreComponents = [
@@ -49,11 +47,6 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
     } else if (component === 'p-text') {
       imports = imports.replace(/, TextSize/, '');
     }
-
-    // add spacing imports
-    imports += ['import type { Spacing }', 'import { getPaddingStyles }']
-      .map((imp, i) => `${i === 0 ? '\n' : ''}${imp} from '../../spacing';`)
-      .join('\n');
 
     // when component is nested we need to fix relative imports
     if (this.shouldGenerateFolderPerComponent(component)) {
@@ -110,6 +103,11 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
     // remove BreakpointCustomizable types since designers can't use JSON
     props = props.replace(/BreakpointCustomizable<(.*)>/g, '$1');
 
+
+    // hidden uxpin props which allows updating property from library level in uxpin editor
+    props = addProp(props, '/** @uxpinignoreprop */ \n  uxpinOnChange: (prevValue: any, nextValue: any, propertyName: string) => void;');
+
+
     // remove useless props
     if (component === 'p-marque') {
       props = removeProp(props, 'href');
@@ -146,10 +144,6 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
       props = addUxPinBindAnnotation(props, 'activeTabIndex', 'onTabChange', 'activeTabIndex');
     }
 
-    // add spacing props to every component
-    const spacings = this.spacingProps.map((x) => `${x}?: Spacing;`).join('\n  ');
-    props = props.replace(/BaseProps & \{\n/, `$&  ${spacings}\n`);
-
     return props;
   }
 
@@ -161,16 +155,6 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
       .replace(/\s+class.*/, ''); // remove class mapping via useMergedClass since it is useless
 
     cleanedComponent = this.insertComponentAnnotation(cleanedComponent, component);
-
-    // destructure spacing props
-    const spacings = this.spacingProps.join(', ');
-    cleanedComponent = cleanedComponent.replace(/(\.\.\.rest)/, `${spacings}, $1`);
-
-    // build inline style prop
-    cleanedComponent = cleanedComponent.replace(
-      /(\.\.\.rest,\n)/,
-      `$1      style: getPaddingStyles({ ${spacings} }),\n`
-    );
 
     // add default children for components that need it
     if (cleanedComponent.includes('PropsWithChildren')) {
@@ -240,6 +224,21 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
     } else if (component === 'p-marque') {
       cleanedComponent = removeDestructuredProp(cleanedComponent, 'href');
       cleanedComponent = removeDestructuredProp(cleanedComponent, 'target');
+    }
+
+    if (component === 'p-pin-code') {
+      cleanedComponent = cleanedComponent.replace(
+          'useEventCallback(elementRef, \'update\', onUpdate as any);',
+          [
+              'const eventCallback = (e:Event) => {',
+              '       rest.uxpinOnChange(value, (e as CustomEvent<PinCodeUpdateEventDetail>).detail.value, \'value\');',
+              '       if (onUpdate) {',
+              '         onUpdate(e as CustomEvent<PinCodeUpdateEventDetail>);',
+              '       }',
+              '    }',
+              '    useEventCallback(elementRef, \'update\', eventCallback);',
+          ].join('\n')
+      )
     }
 
     // cast BreakpointCustomizable default prop values to any because BreakpointCustomizable types are removed for uxpin
@@ -329,7 +328,7 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
         props: { heading: 'Heading', open: true },
         children: [
           '<Text uxpId="modal-text">Some Content</Text>',
-          '<ButtonGroup uxpId="modal-button-group" spacingTop={32}>',
+          '<ButtonGroup uxpId="modal-button-group" >',
           ...[
             '<Button uxpId="modal-button-1" children="Save" />',
             '<Button uxpId="modal-button-2" variant="tertiary" children="Close" />',
@@ -586,6 +585,7 @@ export default <${formComponentName} ${stringifiedProps} />;
     webpackConfig: 'webpack.config.js',
   },
   name: 'Porsche Design System',
+  settings: { useUXPinProps: true, useFitToContentAsDefault: true },
 };`;
 
     return { name: 'uxpin.config.js', relativePath: '../../..', content };